Allow NixpkgsPlatformGroup, use distribution-nixpkgs for parsing
This change no longer exclusively produces `NixpkgsPlatformSinge`s. The
main hurdle for this was parsing: We need to be able to denote platform
groups in the yaml configuration somehow. nixpkgsPlatformFromString
which was added in <https://github.com/NixOS/distribution-nixpkgs/pull/19>
fills in this gap: Any string prefixed with either `platforms.` or
`lib.platforms.` will be interpreted as a platform group, all other
strings will (roughly) be treated as before.

Since the platform parsing capabilities of distribution-nixpkgs no
longer rely on a hard-coded list of known platforms, we can also reuse
it to parse the `--platform` flag of `hackage2nix`. This is a no
brainer, since we expect nixpkgs style `system` tuples anyways and the
parsing is then consistent with the parsing of the hackage2nix
configuration file(s).

`cabal2nix`'s parsing of the (unfittingly named) `--system` flag is not
changed for backwards compatibility, since it emulates the
`GHC_CONVERT_OS` and `GHC_CONVERT_ARCH` flags and does other
preprocessing. This can probably be changed without too grave
consequences, but should be deliberated separately.